Django/D05/requirements.txt
asgiref==3.5.2
Django==4.0.4
psycopg2-binary==2.8.6
sqlparse==0.4.2
